feat(web): move intake and booking to single-page landing flow
