+++ pages/index.js
@@ -1,11 +1,10 @@
 import { useRef, useState } from 'react'
 
 import BenchmarkSection from '@components/BenchmarkSection'
+import ContactBookingSection from '@components/ContactBookingSection'
 import Developers from '@components/Developers'
 import DevToolsSection from '@components/DevToolsSection'
 import EcosystemSection from '@components/EcosystemSection'
-import EmailForm from '@components/EmailForm'
-import FeedbackForm from '@components/FeedbackForm'
 import Footer from '@components/Footer'
 import IndustriesGrid from '@components/IndustriesGrid'
 import MastHead from '@components/MastHead'
@@ -48,7 +47,7 @@ export default function Home() {
                 sectionRef={sectionRef}
             />
             {/* <SocialSection /> */} {/* Temporarily disabled - feeds not working */}
-            <EmailForm />
+            <ContactBookingSection />
             <Footer />
         </div>
     )
